Appears to solve #75 by modularizing parser
The work is not done, however, until more things are pushed into their own files, and overall structure is further considered
