+++ src/parse-helpers.js
@@ -1,0 +1,68 @@
+var P = require("parsimmon");
+
+var _ = require("./parse/whitespace")(null);
+
+function cons(x, xs) {
+    if (arguments.length === 1) {
+        return cons.bind(null, x);
+    }
+    return [x].concat(xs);
+}
+
+function wrap(left, middle, right) {
+    return P.string(left)
+        .then(_)
+        .then(middle)
+        .skip(_)
+        .skip(P.string(right));
+}
+
+function spaced(par) {
+    return _.then(par).skip(_);
+}
+
+function word(str) {
+    return P.string(str).skip(_);
+}
+
+function spread(f) {
+    return function(xs) {
+        return f.apply(null, xs);
+    };
+}
+
+function list0(sep, par) {
+    return list1(sep, par).or(P.of([]));
+}
+
+function list1(sep, par) {
+    if (!sep.then) {
+        console.error(sep.toString());
+        throw new Error("no sep.then");
+    }
+    var more = sep.then(par).many();
+    return par.chain(function(r) {
+        return more.map(function(rs) {
+            return cons(r, rs);
+        });
+    });
+}
+
+function foldLeft(f) {
+    return function(z, xs) {
+        return xs.reduce(function(acc, x) {
+            return f(acc, x);
+        }, z);
+    };
+}
+
+module.exports = {
+    cons: cons,
+    wrap: wrap,
+    spaced: spaced,
+    word: word,
+    spread: spread,
+    list0: list0,
+    list1: list1,
+    foldLeft: foldLeft
+};
